split up authManagement tests into sub files for easier reading
